+++ src/pages/Withdraw/styled.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+// eslint-disable-next-line import/prefer-default-export
+export const Link = styled.a`
+  color: #db3a3d;
+`;
