+++ src/pages/index.ts
@@ -1,0 +1,7 @@
+import PoolsPage from './Pools';
+import DepositPage from './Deposit';
+import WithdrawPage from './Withdraw';
+
+export const Pools = PoolsPage;
+export const Deposit = DepositPage;
+export const Withdraw = WithdrawPage;
